Fixed a disaster I accidentally created (I apparently cannot remember my trig functions sometimes) Fixed limiter and duration to protect ears (the sound is supposed to rise in amplitude, butthis means that it will become too harsh in long durations) This project is done until it is graded and then I'm going to work on it again (this time as a "fun" experiment)

+++ Simple Sound Generator-CIS 1202 Final (Project)/Sound.h
@@ -36,8 +36,8 @@ public:
 class Sine {
 public:
 	//setup for the sine sound
-	const int maxAMPLITUDE = (pow(16, 2) / 2) - 1;//the number of bits to the second power divided by 2 (to factor both negative and positive values) while subtracting 1 to factor in zero
-	float duration;//how long the sound will be (user will input the duration in seconds)
+	const double maxAMPLITUDE = (pow(16, 2) / 2) - 1;//the number of bits to the second power divided by 2 (to factor both negative and positive values) while subtracting 1 to factor in zero
+	double duration;//how long the sound will be (user will input the duration in seconds)
 	double frequencyInput;
 
 };
